refactor: file style changes
