+++ jonathanearl.io/src/home/index.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function Home() {
+    return (<div>HOME HOME ON THE RANGE</div>);
+}
